Phone-based auth + decimal (gram) quantities
Login/register now key on phoneNumber instead of email:
- Login: look up by PhoneNumber; error "Invalid mobile number or password"
- Register: email optional (User.Email now nullable), PhoneNumber unique
  (service check + unique index); JWT email claim added only when present
- 10-digit Indian mobile validation on login/register
- Seed/backfill admin PhoneNumber (SeedAdmin:PhoneNumber, default 9999999999)
  so admin can still sign in under phone-based login

Password reset token lifetime extended 30m -> 24h (custom token, not
DataProtectionTokenProviderOptions which this app does not use).

Cart/order/stock support decimal quantities for gram-based ordering:
- CartItem.Quantity, OrderItem.Quantity, Product.StockQuantity -> decimal(18,3)
- Matching DTOs and inventory endpoint -> decimal; cart validators allow > 0

Adds migration AddPhoneAuthAndDecimalQuantities.

+++ Fruitmarket.API/Controllers/InventoryController.cs
@@ -12,10 +12,10 @@ namespace Fruitmarket.API.Controllers;
 public sealed class InventoryController(IUnitOfWork uow) : ControllerBase
 {
     [HttpPatch("products/{productId:guid}/stock")]
-    public async Task<IActionResult> UpdateStock(Guid productId, [FromBody] int quantity, CancellationToken ct)
+    public async Task<IActionResult> UpdateStock(Guid productId, [FromBody] decimal quantity, CancellationToken ct)
     {
         var product = await uow.Products.GetByIdAsync(productId, ct) ?? throw new ApiException("Product not found", 404);
-        product.StockQuantity = Math.Max(0, quantity);
+        product.StockQuantity = Math.Max(0m, quantity);
         await uow.SaveChangesAsync(ct);
         return NoContent();
     }
