+++ Fruitmarket.Domain/Entities/CartItem.cs
@@ -8,5 +8,6 @@ public sealed class CartItem : BaseEntity
     public Cart? Cart { get; set; }
     public Guid ProductId { get; set; }
     public Product? Product { get; set; }
-    public int Quantity { get; set; }
+    // Decimal to support gram-based ordering (0.25 = 250g, 1.0 = 1kg).
+    public decimal Quantity { get; set; }
 }
